refactor: 컨트롤러 계층 네이밍 변경 및 WasHandlerAdapter 추가
컨트롤러 계층의 구조의 이해를 돕기 위해 네이밍을 변경하였으며, 기존 user에 관한 로직 처리를 조건문으로 분기했던 것을 맞는 Adapter를 호출해주는 HandlerAdapter를 구현하여 Map형식으로 처리함. 이를 위해 이전 UserController 인터페이스와 마찬가지로 WasHandlerAdapter 인터페이스를 구현함.
