+++ src/main/java/util/ViewResolver.java
@@ -1,4 +1,4 @@
-package webserver;
+package util;
 
 import java.io.IOException;
 import java.nio.file.Files;
