+++ src/main/java/webserver/controller/FrontServlet.java
@@ -6,28 +6,32 @@ import model.response.Response;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.HashMap;
+import java.util.Map;
 
 import static model.response.HttpStatusCode.OK;
-import static webserver.ViewResolver.findActualFile;
-import static webserver.ViewResolver.findFilePath;
+import static util.ViewResolver.findActualFile;
+import static util.ViewResolver.findFilePath;
 
 public class FrontServlet {
 
-    private final UserFrontServlet userFrontServlet = new UserFrontServlet();
+    private final Map<String, WasHandlerAdapter> adapterMap = new HashMap<>();
 
+    public FrontServlet() {
+        adapterMap.put("user", new UserHandlerAdapter());
+    }
 
     public void process(Request request, Response response) {
-
         try {
             Path path = findFilePath(request.getUrl());
             response.addHeader("Content-Type", Files.probeContentType(path));
             response.setBody(findActualFile(path));
-
             response.setStatusCode(request.getHttpVersion(), OK);
+
         } catch (IOException e) {
-            if (request.getUrl().contains("user")) {
-                userFrontServlet.process(request, response);
-            }
+            String[] split = request.getUrl().split("/");
+            WasHandlerAdapter wasHandlerAdapter = adapterMap.get(split[1]);
+            wasHandlerAdapter.process(request, response);
         }
     }
 }
