+++ src/main/java/webserver/controller/UserController.java
@@ -2,7 +2,7 @@ package webserver.controller;
 
 import model.request.Request;
 import model.response.Response;
-import webserver.UserService;
+import webserver.service.UserService;
 
 public interface UserController {
     UserService userService = new UserService();
