+++ src/main/java/webserver/controller/UserLoginController.java
@@ -5,7 +5,7 @@ import model.request.Request;
 import model.response.Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.UserService;
+import webserver.service.UserService;
 
 public class UserLoginController implements UserController {
     private static final Logger logger = LoggerFactory.getLogger(UserLoginController.class);
