+++ src/main/java/webserver/controller/WasHandlerAdapter.java
@@ -1,0 +1,10 @@
+package webserver.controller;
+
+import model.request.Request;
+import model.response.Response;
+
+public interface WasHandlerAdapter {
+
+    void process(Request request, Response response);
+
+}
